technically correct upload_path in response project and filenames for uploaded parm7/rst7
GEMS may just want full pathnames for the parm7/rst7 in the project, and we might do away with upload path as GEMS may not understand this. It's the website's job to make requests with the valid pathings.

For now, cleanest hack I can come up with to ensure response project files don't contain full paths like before.